Group `Text` body examples
